# api/openapi.json
{
  "components": {
    "headers": {
      "PanoApiLevel": {
        "description": "The API level this Pano implements.",
        "schema": {
          "type": "integer"
        }
      }
    },
    "responses": {
      "Error400": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Bad request: BAD_REQUEST",
        "x-pano-error-codes": [
          "BAD_REQUEST"
        ]
      },
...
      "Error429": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Too many requests: RATE_LIMIT_EXCEEDED",
        "x-pano-error-codes": [
          "RATE_LIMIT_EXCEEDED"
        ]
      },
      "Error500": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Internal server error: INTERNAL_SERVER_ERROR",
        "x-pano-error-codes": [
          "INTERNAL_SERVER_ERROR"
        ]
      },
      "Error503": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Service unavailable: MAINTENANCE_MODE_ENABLED",
        "x-pano-error-codes": [
          "MAINTENANCE_MODE_ENABLED"
        ]
      }
    },
    "schemas": {
      "Error": {
        "description": "The body of every non-2xx answer.",
        "properties": {
          "error": {
            "properties": {
              "code": {
                "description": "A stable, declared code.",
                "pattern": "^[A-Z][A-Z0-9_]*$",
                "type": "string"
              },
              "details": {
                "additionalProperties": true,
                "description": "Extra facts about the failure.",
                "type": "object"
              },
              "fields": {
                "additionalProperties": {
                  "type": "string"
                },
                "description": "Field name to code, for a rejected form.",
                "type": "object"
              },
              "message": {
                "description": "Optional English text.",
                "type": "string"
              }
            },
            "required": [
              "code"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "A site session token, sent by a server-side front-end that holds a front-end key. Not valid on the panel API.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "The session cookie set by login. A mutation also repeats the `pano_csrf_token` cookie in the `X-CSRF-Token` header. Behind a plain-http origin the cookie names end in `_http`.",
        "in": "cookie",
        "name": "pano_auth_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "The public API of the plugin pano-plugin-market, served below /api/v1/plugins/pano-plugin-market.",
    "summary": "Pano plugin API: pano-plugin-market",
    "title": "Pano plugin API: pano-plugin-market",
    "version": "1",
    "x-pano-api-level": 1,
    "x-pano-plugin-id": "pano-plugin-market",
    "x-pano-scope": "plugin"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/orders/{publicId}/pay": {
      "post": {
        "operationId": "PayOrder",
        "parameters": [
          {
            "in": "path",
            "name": "publicId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "payment": {
                      "properties": {
                        "kind": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "kind"
                      ],
                      "type": [
                        "object",
                        "null"
                      ]
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Success",
            "headers": {
              "Pano-Api-Level": {
                "$ref": "#/components/headers/PanoApiLevel"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not found: NOT_FOUND",
            "x-pano-error-codes": [
              "NOT_FOUND"
            ]
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Conflict: ORDER_NOT_PAYABLE",
            "x-pano-error-codes": [
              "ORDER_NOT_PAYABLE"
            ]
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Service unavailable: MAINTENANCE_MODE_ENABLED, STORE_DISABLED, STORE_UNAVAILABLE",
            "x-pano-error-codes": [
              "MAINTENANCE_MODE_ENABLED",
              "STORE_DISABLED",
              "STORE_UNAVAILABLE"
            ]
          }
        },
        "summary": "Starts a payment for a pending order with the chosen method.",
        "tags": [
          "orders"
        ],
        "x-pano-stability": "public"
      }
    },
...
    "/payments/attempts/{attemptToken}/page": {
      "get": {
        "operationId": "GetPaymentAttemptPage",
        "parameters": [
          {
            "in": "path",
            "name": "attemptToken",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Success",
            "headers": {
              "Pano-Api-Level": {
                "$ref": "#/components/headers/PanoApiLevel"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not found: NOT_FOUND",
            "x-pano-error-codes": [
              "NOT_FOUND"
            ]
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "summary": "The hosted page of a payment attempt, for providers that show instructions.",
        "tags": [
          "payments"
        ],
        "x-pano-stability": "public"
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/plugins/pano-plugin-market"
    }
  ],
...
}